refactor(basic): point 렌더 함수 view폴더로 이동
- point관련 뷰 그려주는 함수가 domain폴더에 있었는데 view폴더로 이동
